Machine Vertex for MunichMotorDevice
