+++ spynnaker/pyNN/external_devices_models/machine_munich_motor_device.py
@@ -1,0 +1,45 @@
+# Copyright (c) 2017-2019 The University of Manchester
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+from pacman.model.graphs.machine import MachineVertex
+from spinn_utilities.overrides import overrides
+from spinn_front_end_common.abstract_models import (
+    AbstractHasAssociatedBinary)
+from spinn_front_end_common.utilities.utility_objs import ExecutableType
+
+
+class MachineMunichMotorDevice(MachineVertex, AbstractHasAssociatedBinary):
+    """ An Omnibot motor control device. This has a real vertex and an \
+        external device vertex.
+    """
+    def __init__(self, resources, label=None, constraints=None,
+                 app_vertex=None, vertex_slice=None):
+        super(MachineMunichMotorDevice, self).__init__(
+            label=label, constraints=constraints, app_vertex=app_vertex,
+            vertex_slice=vertex_slice)
+        self._resources = resources
+
+    @property
+    @overrides(MachineVertex.resources_required)
+    def resources_required(self):
+        return self._resources
+
+    @overrides(AbstractHasAssociatedBinary.get_binary_file_name)
+    def get_binary_file_name(self):
+        return "robot_motor_control.aplx"
+
+    @overrides(AbstractHasAssociatedBinary.get_binary_start_type)
+    def get_binary_start_type(self):
+        return ExecutableType.USES_SIMULATION_INTERFACE
